TASK 8
USEEFFECT SOLVED PROPBLEM
When you pass an object as a dependency, React uses reference equality to determine if the object has changed between renders. If the object reference is different from the previous render, the effect will run. This is because React doesn't perform a deep comparison of object properties.